Get the specs to run
If sass or coffee-script is not installed, or execjs can not find
a javascript runtime, skip the assets test with an appropriate
message.

Don't use eq in the specs, as it isn't available in all versions
of RSpec Roda supports (1.3+).

Remove check for external assets, since the support was removed.

Fix spec after removal of :concat_name option.

A spec still fails if yuicompressor is not installed.