Remove assets_unique_id method, store in assets_opts hash
When compiling the assets, calculate the unique id before writing
to the path, and include the unique id in the path.

Have compile_assets return a hash with the unique ids for each type,
useful if precompiling the assets.

Try to use the matching path in Roda#assets.

There is some folder path munging with splitting on / and -, I'm
guessing this affects that.  But the previous code was busted in
regards to folders anyway as assets_unique_id only took a type,
not a folder.