More changes
Fix compile_assets to use the correct key for unique_ids.

Fix Roda#assets so it uses correct hash and local variable name,
and works correctly when folder is an array.

Simplify compile_assets and compile_process_files.

Allow compile_assets to be called with a single asset type.

Require digest/sha1 in compile_process_files, since it is needed there.

By default, write compiled file to public directory in the path that
Roda#assets would reference the file.  This allows the compiled file
to be served directly by the front end webserver.

Use app instead of compiled.roda.assets as the compiled file name.

If :unique_ids option is used, or compile_assets is called, set
:compiled to true.

Work with a single file name in addition to an array of file names.

Update the documentation.