Add the symbolized_params plugin
This is an alternative plugin to indifferent_params. Using indifferent
parameters is always a bit awkward, because it makes you believe you
have a Hash with symbolized keys, until you try to use something like
`Hash#fetch`, or when you try to pass the params as keyword arguments.

This plugin is not recommended for Ruby versions lower than 2.2.0,
because only in this version Symbol GC was introduced.