Allow overriding exception page assets in code
While the exception page is able to use static assets, this becomes
a hassle when dealing with writing Roda extensions or app frameworks
based on Roda. For example, Bridgetown uses Roda and configures the
exception page plugin with Bridgetown branding ... but does so with
a monkey-patch to the `Roda::RodaPlugins::ExceptionPage` class.

This change gives an affordance for overriding the CSS and JavaScript
content of the exception page assets, falling back to the default
behavior when not configured.